+++ src/annoy.h
@@ -12,7 +12,7 @@ class AnnoyIndexWrapper : public Napi::ObjectWrap<AnnoyIndexWrapper>
 public:
   static Napi::Object Init(Napi::Env env, Napi::Object exports);
   AnnoyIndexWrapper(const Napi::CallbackInfo &info);
-  AnnoyIndexInterface<int, double, uint32_t> *t;
+  AnnoyIndexInterface<int, float, uint32_t> *t;
 
 private:
   Napi::Value addItem(const Napi::CallbackInfo &info);
