+++ test/test.ts
@@ -3,7 +3,7 @@ import Annoy, { Metric } from "../dist";
 const annoy = new Annoy(128, Metric.ANGULAR);
 
 for (let i = 0; i < 1000; i += 1) {
-  const vector = new Float64Array(128);
+  const vector = new Float32Array(128);
   for (let v = 0; v < 128; v += 1) {
     vector[v] = Math.random();
   }
